Refactorizando la capa de Dominio
