--- src/main/java/com/ferrotodo/domain/GestionDePersonalImpl.java
@@ -1,7 +1,0 @@
-package com.ferrotodo.domain;
-
-import com.ferrotodo.domain.negocio.GestionDePersonal;
-
-public class GestionDePersonalImpl implements GestionDePersonal {
-
-}
