--- src/main/java/com/ferrotodo/domain/InventarioImpl.java
@@ -1,22 +1,0 @@
-package com.ferrotodo.domain;
-
-import com.ferrotodo.domain.entities.Producto;
-import com.ferrotodo.domain.negocio.Inventario;
-
-public class InventarioImpl implements Inventario {
-    @Override
-    public void agregarProducto(Producto producto, int cantidad) {
-
-    }
-
-    @Override
-    public void reducirProducto(Producto producto, int cantidad) {
-
-    }
-
-    @Override
-    public int obtenerCantidad(Producto producto) {
-        return 0;
-    }
-
-}
