--- src/main/java/com/ferrotodo/domain/TransaccionImpl.java
@@ -1,7 +1,0 @@
-package com.ferrotodo.domain;
-
-import com.ferrotodo.domain.negocio.Transaccion;
-
-public class TransaccionImpl implements Transaccion {
-
-}
