--- src/main/java/com/ferrotodo/domain/negocio/GestionDePersonal.java
@@ -1,5 +1,0 @@
-package com.ferrotodo.domain.negocio;
-
-public interface GestionDePersonal {
-
-}
