+++ src/main/java/com/ferrotodo/domain/negocio/GestionDePersonalImpl.java
@@ -1,0 +1,7 @@
+package com.ferrotodo.domain.negocio;
+
+import com.ferrotodo.domain.ports.input.Gestionable;
+
+public class GestionDePersonalImpl implements Gestionable {
+
+}
