+++ src/main/java/com/ferrotodo/domain/negocio/TransaccionImpl.java
@@ -1,0 +1,7 @@
+package com.ferrotodo.domain.negocio;
+
+import com.ferrotodo.domain.ports.input.Transaccionable;
+
+public class TransaccionImpl implements Transaccionable {
+
+}
