+++ src/main/java/com/ferrotodo/domain/ports/input/Gestionable.java
@@ -1,0 +1,5 @@
+package com.ferrotodo.domain.ports.input;
+
+public interface Gestionable {
+
+}
