+++ src/main/java/com/ferrotodo/domain/ports/input/Transaccionable.java
@@ -1,0 +1,5 @@
+package com.ferrotodo.domain.ports.input;
+
+public interface Transaccionable {
+
+}
